Refactoring openhtf/__init__.py moving phase related objects into phase_descriptor.py and test related objects into test_descriptor.py.
Cleaned up many low-hanging linting errors (but not all errors in these files).

Note:  I'm going to try using the new copybara change request work flow I setup to sync this change to github/openhtf.  I think the best workflow would be to wait until we receive a LGTM internally.  Then sync to github openhtf (get a rubber stamp LGTM from same reviewer) and submit there.  Then pull that change in.

PiperOrigin-RevId: 192494082
